requirements.txt
Django>=4.2
djangorestframework>=3.14
channels>=4.0
asgiref>=3.6

Pillow>=9.0
dj-database-url>=1.0
firebase-admin>=6.0
